# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LEVH API",
    "description": "Local-first memory layer for AI agents and humans",
    "version": "2.32.0"
  },
  "paths": {
...
    "/api/v1/agents/connect": {
      "post": {
        "summary": "Agent Connect",
        "description": "Record an agent connecting to LEVH.",
        "operationId": "agent_connect_api_agents_connect_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentConnectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentConnectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/context": {
      "get": {
        "summary": "Get Context",
        "operationId": "get_context_api_context_get_v1",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Id"
            }
          },
          {
            "name": "project",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project"
            }
          },
          {
            "name": "max_tokens",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 4000,
              "title": "Max Tokens"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/context-file": {
      "post": {
        "summary": "Generate Context File",
        "description": "Generate a CLAUDE.md / .cursorrules style context file from memories.",
        "operationId": "generate_context_file_api_context_file_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/connectors/import": {
      "post": {
        "summary": "Connector Import",
        "description": "Import data from an external app via connector.",
        "operationId": "connector_import_api_connectors_import_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/connectors/sync": {
      "post": {
        "summary": "Connector Sync",
        "description": "Connector v2 ingest: fetch, then route items through the admission\ngate (dedupe + secret redaction), with incremental sync bookkeeping.",
        "operationId": "connector_sync_api_connectors_sync_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/connectors/upload": {
      "post": {
        "summary": "Connector Upload",
        "description": "Store an uploaded file locally and return the path to import from.",
        "operationId": "connector_upload_api_connectors_upload_post_v1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadedFileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/connectors/{name}/config": {
      "get": {
        "summary": "Get Connector Config",
        "description": "Get required config fields for a connector.",
        "operationId": "get_connector_config_api_connectors__name__config_get_v1",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentConnectRequest": {
        "properties": {
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "project": {
            "type": "string",
            "title": "Project",
            "default": ""
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "agent_name"
        ],
        "title": "AgentConnectRequest"
      },
      "AgentConnectResponse": {
        "properties": {
          "agent_session_id": {
            "type": "string",
            "title": "Agent Session Id"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "display": {
            "type": "string",
            "title": "Display"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "connected_at": {
            "type": "string",
            "title": "Connected At"
          }
        },
        "type": "object",
        "required": [
          "agent_session_id",
          "agent_name",
          "display",
          "icon",
          "connected_at"
        ],
        "title": "AgentConnectResponse",
        "description": "``POST /api/agents/connect`` - the created agent-session record."
      },
...
      "ConnectorConfigResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "required_config_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Config Keys"
          },
          "help": {
            "type": "string",
            "title": "Help"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "help"
        ],
        "title": "ConnectorConfigResponse",
        "description": "``GET /api/connectors/{name}/config`` adds the help text."
      },
...
      "ConnectorRequest": {
        "properties": {
          "connector": {
            "type": "string",
            "title": "Connector"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "default": {}
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "use_gate": {
            "type": "boolean",
            "title": "Use Gate",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "connector"
        ],
        "title": "ConnectorRequest"
      },
      "ConnectorSyncResponse": {
        "properties": {
          "connector": {
            "type": "string",
            "title": "Connector"
          },
          "fetched": {
            "type": "integer",
            "title": "Fetched"
          },
          "stored": {
            "type": "integer",
            "title": "Stored"
          },
          "redacted": {
            "type": "integer",
            "title": "Redacted"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          },
          "held": {
            "type": "integer",
            "title": "Held"
          },
          "errors": {
            "type": "integer",
            "title": "Errors"
          },
          "source_key": {
            "type": "string",
            "title": "Source Key"
          },
          "last_synced_at": {
            "type": "string",
            "title": "Last Synced At"
          }
        },
        "type": "object",
        "required": [
          "connector",
          "fetched",
          "stored",
          "redacted",
          "duplicates",
          "held",
          "errors",
          "source_key",
          "last_synced_at"
        ],
        "title": "ConnectorSyncResponse",
        "description": "The shared ingest report of ``/api/connectors/import`` and\n``/api/connectors/sync``."
      },
...
      "ConnectorUploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_b64": {
            "type": "string",
            "title": "Content B64"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content_b64"
        ],
        "title": "ConnectorUploadRequest"
      },
...
      "ContextFileRequest": {
        "properties": {
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "style": {
            "type": "string",
            "title": "Style",
            "default": "claude"
          }
        },
        "type": "object",
        "title": "ContextFileRequest"
      },
      "ContextFileResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content"
        ],
        "title": "ContextFileResponse"
      },
      "ContextResponse": {
        "properties": {
          "context": {
            "type": "string",
            "title": "Context"
          },
          "chars": {
            "type": "integer",
            "title": "Chars"
          }
        },
        "type": "object",
        "required": [
          "context",
          "chars"
        ],
        "title": "ContextResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadedFileResponse": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "bytes": {
            "type": "integer",
            "title": "Bytes"
          }
        },
        "type": "object",
        "required": [
          "path",
          "filename",
          "bytes"
        ],
        "title": "UploadedFileResponse",
        "description": "Shared by the upload endpoints: where the bytes landed on this\nmachine, and the display name the caller chose (never the stored path)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}